Move different events to separate files
