+++ source/device/switch.hpp
@@ -2,7 +2,7 @@
 
 #include "device/interfaces/i_switch.hpp"
 #include "device/scheduling_module.hpp"
-#include "event.hpp"
+#include "event/process.hpp"
 
 namespace sim {
