+++ source/event/stop.hpp
@@ -1,0 +1,16 @@
+#pragma once
+#include "event.hpp"
+
+namespace sim {
+
+/**
+ * Stop simulation and clear all events remaining in the Scheduler
+ */
+class Stop : public Event {
+public:
+    Stop(Time a_time);
+    virtual ~Stop() = default;
+    void operator()() final;
+};
+
+}
