+++ source/event/tcp_metric.cpp
@@ -1,0 +1,24 @@
+#include "flow/interfaces/i_tcp_flow.hpp"
+#include "metrics/metrics_collector.hpp"
+#include "scheduler.hpp"
+#include "tcp_metric.hpp"
+
+namespace sim {
+
+TcpMetric::TcpMetric(Time a_time, std::weak_ptr<ITcpFlow> a_flow)
+    : Event(a_time), m_flow(a_flow) {};
+
+void TcpMetric::operator()() {
+    if (m_flow.expired()) {
+        return;
+    }
+
+    auto flow = m_flow.lock();
+
+    double cwnd = flow->get_cwnd();
+    MetricsCollector::get_instance().add_cwnd(flow->get_id(), m_time, cwnd);
+
+    Scheduler::get_instance().add<TcpMetric>(m_time + DELAY, m_flow);
+}
+
+}  // namespace sim
