+++ source/flow/tcp_flow.hpp
@@ -2,6 +2,7 @@
 
 #include "device/interfaces/i_host.hpp"
 #include "flow/interfaces/i_tcp_flow.hpp"
+#include "packet.hpp"
 
 namespace sim {
