+++ source/simulator.hpp
@@ -7,6 +7,8 @@
 
 #include "device/host.hpp"
 #include "device/switch.hpp"
+#include "event/start_flow.hpp"
+#include "event/stop.hpp"
 #include "flow/basic_flow.hpp"
 #include "flow/tcp_flow.hpp"
 #include "link/link.hpp"
